posts-router.js
setup posts-router.js in data with express, a router, router.use, and brought in db.js. I setup router.get/ with status200 and post if sucessful with status 500 if data couldn't be retrieved. Setup router.get('/:id') that returns status200 and post if successful, status400 if the post could not be found by the id, and status 500 if the data couldn't be retrieved. setup router.post('/' that required title and contents with status 400 if either are missing or not valid, status200 with post if successful, status 404 if the post couldn't be found by the id, and status 500 if the list could not update with new post