delete in posts-router.js
setup router.delete('/:id'), if the post is removed successfully, we find the next successful post to return with status200, if post was not deleted, status 404 due to post could not be found with id. If the removal is unsuccessful all together, status500 with message that the post could not be deleted.