post comments
to post a comment, we require text in the body, if no text, status 400. Then, we find post by id, if we find the post, we insert a comment with text to the post's array. If the comment length is greater than 0, we return status 200 with the comments, otherwise, we get status 404 because we couldn't find a comment with the right id.